+++ providers/gem.rb
@@ -2,7 +2,19 @@ action :install do
   g = gem_package new_resource.name do
     gem_binary Sensu::Helpers.gem_binary
     version new_resource.version
+    source  new_resource.source
     options new_resource.options
+  end
+
+  new_resource.updated_by_last_action(g.updated_by_last_action?)
+end
+
+action :upgrade do
+  g = gem_package new_resource.name do
+    gem_binary Sensu::Helpers.gem_binary
+    version new_resource.version
+    options new_resource.options
+    action :upgrade
   end
 
   new_resource.updated_by_last_action(g.updated_by_last_action?)