+++ resources/gem.rb
@@ -1,6 +1,7 @@
-actions :install, :remove
+actions :install, :upgrade, :remove
 
 attribute :version, :kind_of => String
+attribute :source,  :kind_of => String
 attribute :options, :kind_of => [String, Hash]
 
 def initialize(*args)