+++ test/integration/default/serverspec/default_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+
+describe file("/etc/sensu/config.json") do
+  its(:content) { should match /"name": "test"/ }
+  its(:content) { should_not match /"id": / }
+  it { should be_grouped_into('nogroup') }
+end