+++ test/integration/enterprise/serverspec/enterprise_spec.rb
@@ -1,0 +1,7 @@
+require 'spec_helper'
+require 'service_dependency_helper'
+
+describe service("sensu-enterprise") do
+  it { should be_enabled }
+  it { should be_running }
+end