+++ test/integration/helpers/serverspec/spec_helper.rb
@@ -1,0 +1,8 @@
+require "net/http"
+require 'serverspec'
+require "uri"
+
+set :backend, :exec
+set :path, "/bin:/usr/bin:/sbin:/usr/sbin"
+
+puts "os: #{os}"