+++ CHANGELOG.md
@@ -2,11 +2,23 @@
 
 This file is used to track changes made in each version of the sensu cookbook.
 
-## Unreleased
+## 2.12.0 - 2016-03-14
 
 ### Project changes
 
 The Sensu cookbook project has adopted a new contribution workflow and a new code of conduct policy. Please see the relevant documents in repo for details.
+
+### Behavior changes
+
+* Values for `owner` and `group` properties on `sensu_json_file` resources now default to lazy evaluation of node attributes `sensu.admin_user` and `sensu.group` respectively. (#426)
+
+* Data bags remain default source of SSL certificates, but are now optional:
+
+	With the addition of Sensu state helpers in #410 recipes which access credentials via data bags (i.e. `default`, `rabbitmq` and `enterprise` recipes) have been updated to make these data bag items optional.
+
+	Please see the readme and integration test suite for examples of using these helpers.
+
+* Testing notes have been added in `TESTING.md` to describe some of the platform/suite combinations which are disabled or otherwise require special configuration.
 
 ### Features
 
@@ -15,6 +27,14 @@ Added [ChefSpec](https://github.com/sethvargo/chefspec) test coverage for the fo
 * `default` and `client_service` recipes
 * `sensu_gem` LWRP
 * Sensu::Helpers library `#select_attributes` and `#gem_binary`  methods
+
+Added source attribute to sensu_gem resource
+Added upgrade action to sensu_gem resource
+Added helpers for storing key/value pairs which persist for duration of the Chef run
+
+### Fixes
+
+Allow "standard" as a value of type attribute on `sensu_check` resources, [as described in Sensu documentation](https://sensuapp.org/docs/0.21/checks).
 
 ## 2.11.0 - 2015-11-11